requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
pydantic>=2.7,<3.0
pyyaml>=6.0,<7.0
scikit-learn>=1.5,<2.0
scipy>=1.13,<2.0
pyarrow>=15.0,<18.0
openpyxl>=3.1,<4.0
openai>=1.30,<2.0
jupyter>=1.0,<2.0
notebook>=7.0,<8.0
ipykernel>=6.29,<7.0
pytest>=8.0,<9.0
ruff>=0.6,<1.0
python-dotenv>=1.0,<2.0
